Cleaned up ColorPicker
